In pursuit of a better clarity, I've made the following changes to the converter code.
I've split ConversionRule into an interface and two implementations:
SingleConversionRule and MultiGroupConversionRule. As its name
indicates, SingleConversionRule admits a single group, the entire
matching text, and replaces it with text specified by the developer.

MultiGroupConversion admits multiple groups with a replacement text
specific for each group. I've removed the map associating indexes with
a replacement text. Instead, MultiGroupConversion uses a bounded array
of replacement texts. I believe the result to be somewhat easier to
follow and to understand -- in particular, the replace(Matcher) method
in both implementations.

Note the addition of AllTest and PackageTest enabling the lauch of
all test in one click. Admittedly, there are only two test classes 
for the time being.