Fixed bug 114. Added test case.
If class B extends A, assuming B does not override method x(), the caller
of new B().x() is A and not B

More specifically, XLogger extends LoggerWrapper. When calling xlogger.debug where debug is a method defined in LoggerWrapper not overridden in XLogger, the caller is LoggerWrapper and not XLogger.