Use _ to shorten tag names instead of *.
A common method of enabling logging for a certain tag involves issuing
the setprop command via adb shell.  Unfortunately, it is impossible to
set a property on a key whose name contains a * using this method.  As
such, * is a bad choice for shortening tag names, since it means we
cannot use this method for setting the log level of a tag.