Introduces MDC.putAllCloseable
This feature offers the same functionality as MDC.putCloseable but for
multiple values. And tries to follow the put/putAll signature one can
see on a Map. This would be extremely useful for structured logging when
the appender is configured to use MDC values.

The MDCCloseable class has been modified in such a way that if a single
value is used the consumed memory profile remains the same.