Update version in all pom.xml files from 1.7.22-SNAPSHOT to 1.7.22.1-SNAPSHOT
Why?
Add a fourth level to the project version number scheme to be able to version the forked build artifacts
while also retaining a clear visibility of the base version (1.7.22) of the original.

In a previous commit the version was already changed in the parent pom.xml, but all other modules
were forgotten to update.