Start using maven-git-versioning-extension to virtually calculate project.version for feature branches
Why?
This allows installing and deploying artifacts to local and remote maven repositories
without having to edit pom.xml files in feature branches. Is such case the "project.version"
will be in the form of "$branchName-SNAPSHOT", where "/" characters are replaced with "-".

The version numbers will be still increased in pom.xml files automatically when the
maven-release-plugin prepares a new release (from master) and commits them.