ci: introduce GitHub Actions to run build in the PR
We use Travis CI to run the build to verify PRs, but it does not report the result to PR page.
So it is possible that we merge PR which breaks the build without being noticed.

I cannot identify the reason of this problem, so suggest using GitHub Actions as a workaround.
It can integrate the build result to each pushed commit without configuration.